+++ src/CpuRaytracing/Camera.cpp
@@ -14,11 +14,11 @@
 
 using namespace Raycer;
 
-void Camera::setImagePlaneSize(int width, int height)
+void Camera::setImagePlaneSize(size_t width, size_t height)
 {
 	imagePlaneWidth = (double)(width - 1);
 	imagePlaneHeight = (double)(height - 1);
-	aspectRatio = (double)height / width;
+	aspectRatio = (double)height / (double)width;
 }
 
 void Camera::calculateVariables()
@@ -68,7 +68,7 @@ void Camera::update(double timeStep)
 	calculateVariables();
 }
 
-Ray Camera::getRay(int x, int y) const
+Ray Camera::getRay(size_t x, size_t y) const
 {
 	double dx = ((double)x / imagePlaneWidth) - 0.5;
 	double dy = ((double)y / imagePlaneHeight) - 0.5;
