+++ src/Rendering/Image.h
@@ -18,7 +18,7 @@ namespace Raycer
 	public:
 
 		Image();
-		Image(int width, int height);
+		Image(size_t width, size_t height);
 		Image(const Image& image);
 		Image(const Framebuffer& framebuffer);
 		Image(const std::string& fileName);
@@ -28,20 +28,20 @@ namespace Raycer
 
 		void load(const std::string& fileName);
 		void saveAs(const std::string& fileName) const;
-		void setSize(int width, int height);
-		void setPixel(int x, int y, const Color& color);
+		void setSize(size_t width, size_t height);
+		void setPixel(size_t x, size_t y, const Color& color);
 		void swapBytes();
 		void flip();
 
-		int getWidth() const;
-		int getHeight() const;
-		Color getPixel(int x, int y) const;
+		size_t getWidth() const;
+		size_t getHeight() const;
+		Color getPixel(size_t x, size_t y) const;
 		Color getPixel(double s, double t) const;
 
 	private:
 
-		int width = 0;
-		int height = 0;
+		size_t width = 0;
+		size_t height = 0;
 		uint32_t* pixelData = nullptr; // RGBA
 	};
 }
